First pass at Lagrangian filtering framework
This is the first steps toward building a framework for flexible
Lagrangian filtering of ocean model data. The current code is intended
to transparently set up OceanParcels to run the filtering algorithm:
creation of a fieldset encapsulating the input data, generation of a
particle class and kernel to sample fields of interest, and a method
to run the advection forwards and backwards through the filter window.

Not yet implemented is the highpass filtering of the particle data and
output to a final file.

This uses a custom branch of OceanParcels that integrates code from
two pull requests (https://github.com/OceanParcels/parcels/pull/497
and https://github.com/OceanParcels/parcels/pull/605), which provide
faster output during advection by using temporary files, and faster
input during advection by chunking the input fields.