Add custom HDF5 ParticleFile for output
Because the output file is only really temporary (we just need to
store the data for long enough to filter, then discard), a new
LagrangeParticleFile class has been implemented to dump particle
output to a temporary HDF5 file. This should allow for efficient
writing/re-reading of the data for filtering.