Sample initial conditions in C kernel
If a Variable on a Particle specifies a Field for an initial
condition, the spatial interpolation is evaluated in Python. For large
numbers of particles, this makes even initialisation completely
intractable. Instead, we leave off the 'initial' parameter, so sampled
variables are initialised to zero. However, we run a sample-only
kernel for a single timestep, which loads data and interpolates over
all particles in a much more efficient C kernel.