Combine advection in single tmpfile, eagerly filter
This doesn't reduce the disk usage by the temporary file, however the
eager computation of the results means that it gets cleaned up
straight away, so we don't persist a large number of mostly useless
temporary files. This would be more useful with a strategy to dump
results to the output file immediately, rather than gathering them
in-memory.

Closes #9.