Output filtered data to a netCDF file
Using the source files, and the variable/dimension dictionaries, a
netCDF dataset is created to hold the filtered data. The approach I've
taken here is pretty ugly, and makes too many assumptions. Ideally,
I'd like to use xarray to do most of the work, but incremental writes
to a dataset are one of the unsupported operations.