Save the original fieldset grid on LagrangeFilter
When we tell parcels to make our domain periodic, it modifies the
underlying grids to add the halo region. Because we were using these
grids to determine the coordinates at which to seed particles, this
would result in particles being seeded in the halos. While this
doesn't necessarily cause problems, it is unneeded computation that
would have to be removed after filtering.

Instead, we save the meshgrids for particle lat/lon on the
LagrangeFilter object itself, so that particle seeding remains
unchanged independent of parcels' underlying modifications.