Add advection sanity check using advection_step()
This revealed a bug: if the time axis doesn't start at zero, the
velocity data is offset. In this case, there was a single timestep
offset, which was expected given the time axis. I didn't verify
whether this offset would increase if, for example, the time axis
started at 2*dt.

See also #20.