Cut off ends of advected data before filtering
Not sure how I missed this one, but ParticleSet.execute writes data at
the end of advection. Because we always do advection on a multiple of
output timesteps, this means that the last timestep is duplicated in
the output data. I had accounted for this in the test cases, but not
in filter_step itself, so the filtering was doubling up the data on
the ends of the series.