Output filtered data to a netCDF file (#14)
* Output filtered data to a netCDF file

Using the source files, and the variable/dimension dictionaries, a
netCDF dataset is created to hold the filtered data. The approach I've
taken here is pretty ugly, and makes too many assumptions. Ideally,
I'd like to use xarray to do most of the work, but incremental writes
to a dataset are one of the unsupported operations.

* Add simple test case for file creation

This is really just the most basic case, and doesn't cover any tricky
corner cases.

* Fix file creation with per-variable dimensions

This adds a few more test cases. In particular, the staggered test
case caught missing behaviour when variables have their own separate
dimensions (as is the case for staggered variables).

* Restructure netcdf output, handle curvilinear grids

* Add test for curvilinear grids

* Expand wildcards when creating output file

I had been testing wildcard filenames before, and in those cases it
seems like parcels expanded the wildcard when creating the
FieldSet. However, with two wildcards, this wasn't happening. We
manually expand wildcards when querying the input datasets for the
metadata needed to create the output file.

Closes #34.

* Default to not clobbering output files

This adds a flag to filtering and create_out to control whether an
existing output file should be clobbered or not.

Closes #27.

* Support creation of netcdf output file from xarray input

* Handle variables with dimensions in indices dict

For example, a data variable might only have dims ("time", "lat",
"lon"), but a "depth" dimension was specified in indices.

* Handle per-variable indices

* Add test for per-variable indices