Make the output file only use dimensions of the particle grid
Instead of copying dimensions from input variables across to the
output file, we now only reflect the grid on which the particles were
seeded in the output file, accurately locating the output data.

To not rely on the first FieldSet grid as the source of all truth, we
now also defer inertial filter creation, etc.