Update for new Parcels particleset and compiler APIs
Parcels is abstracting away to a more generic compiler interface,
but so we need to update the object we refer to, and also pass it
the include directory so it can find the C interface.

As part of the effort to support multiple particle storage
collections, the ParticleSet now simply refers to a generic
collection as a private member. We need access to its ptype member
to figure out which variables are stored on the particles, so
unfortunately we have to maintain this private reference.