Shuffle HDF5-cached data in a numpy array
Advection was taking a very long time, and it seems that the culprit
is the indexed write of valid particle data. Instead of relying on the
HDF5 library to do this, we first write all data to a numpy array, and
write the entire thing at once to the file.